Move Kolmogorov model test to own file.
